Fixed excessive memory consumption by SnapshotInfo.
