+++ cpvc/Core/Core.cs
@@ -59,7 +59,7 @@ namespace CPvC
         public BeginVSyncDelegate BeginVSync { get; set; }
         public IdleRequestDelegate IdleRequest { get; set; }
 
-        public AudioBuffer AudioBuffer
+        public CircularAudioBuffer AudioBuffer
         {
             get
             {
@@ -70,7 +70,7 @@ namespace CPvC
         private AutoResetEvent _audioReady;
         private AutoResetEvent _requestQueueNonEmpty;
 
-        private AudioBuffer _audioBuffer;
+        private CircularAudioBuffer _audioBuffer;
 
         private SynchronizationContext _syncContext;
 
@@ -103,7 +103,7 @@ namespace CPvC
             _audioReady = new AutoResetEvent(true);
             _requestQueueNonEmpty = new AutoResetEvent(false);
 
-            _audioBuffer = new AudioBuffer();
+            _audioBuffer = new CircularAudioBuffer();
 
             // Ensure any OnPropChanged calls are executed on the main thread. There may be a better way of
             // doing this, such as wrapping add and remove for Command.CanExecuteChanged in a lambda that
